Moved kubevirt_vmi_number_of_outdated_after_upgrade
Refactored kubevirt_vmi_number_of_outdated_after_upgrade test
to test_upgrade_iuo.
