+++ tests/conftest.py
@@ -3032,23 +3032,3 @@ def nmstate_namespace(admin_client):
 @pytest.fixture()
 def kubevirt_resource(admin_client, hco_namespace):
     return get_hyperconverged_kubevirt(admin_client=admin_client, hco_namespace=hco_namespace)
-
-
-@pytest.fixture(scope="session")
-def running_vm_with_bridge(
-    unprivileged_client,
-    upgrade_namespace_scope_session,
-    upgrade_br1test_nad,
-):
-    name = "vm-bridge-connected"
-    with VirtualMachineForTests(
-        name=name,
-        namespace=upgrade_namespace_scope_session.name,
-        networks={upgrade_br1test_nad.name: upgrade_br1test_nad.name},
-        interfaces=[upgrade_br1test_nad.name],
-        client=unprivileged_client,
-        body=fedora_vm_body(name=name),
-        eviction_strategy=ES_NONE,
-    ) as vm:
-        running_vm(vm=vm, wait_for_cloud_init=True)
-        yield vm
