+++ tests/install_upgrade_operators/product_upgrade/test_upgrade_iuo.py
@@ -6,6 +6,7 @@ from tests.install_upgrade_operators.product_upgrade.utils import (
     process_alerts_fired_during_upgrade,
     verify_nodes_labels_after_upgrade,
     verify_nodes_taints_after_upgrade,
+    wait_for_greater_than_zero_metric_value,
 )
 from tests.upgrade_params import (
     IMAGE_UPDATE_AFTER_UPGRADE_NODE_ID,
@@ -73,4 +74,16 @@ class TestUpgradeIUO:
             nodes=nodes,
             nodes_labels_before_upgrade=nodes_labels_before_upgrade,
             cnv_upgrade=cnv_upgrade,
+        )
+
+    @pytest.mark.polarion("CNV-11758")
+    @pytest.mark.order(before=IMAGE_UPDATE_AFTER_UPGRADE_NODE_ID, after=IUO_CNV_ALERT_ORDERING_NODE_ID)
+    @pytest.mark.dependency(
+        depends=[IUO_UPGRADE_TEST_DEPENDENCY_NODE_ID],
+        scope=DEPENDENCY_SCOPE_SESSION,
+    )
+    def test_metric_kubevirt_vmi_number_of_outdated_after_upgrade(self, prometheus):
+        wait_for_greater_than_zero_metric_value(
+            prometheus=prometheus,
+            metric_name="kubevirt_vmi_number_of_outdated",
         )
