+++ tests/install_upgrade_operators/product_upgrade/utils.py
@@ -29,6 +29,7 @@ from utilities.constants import (
     FIRING_STATE,
     HCO_CATALOG_SOURCE,
     IMAGE_CRON_STR,
+    TIMEOUT_5MIN,
     TIMEOUT_5SEC,
     TIMEOUT_10MIN,
     TIMEOUT_10SEC,
@@ -50,6 +51,7 @@ from utilities.infra import (
     wait_for_consistent_resource_conditions,
     wait_for_version_explorer_response,
 )
+from utilities.monitoring import get_metrics_value
 from utilities.operator import (
     approve_install_plan,
     get_hco_version_name,
@@ -716,3 +718,21 @@ def wait_for_odf_update(target_version: str) -> None:
         if not sample:
             return
         LOGGER.info(f"Following odf csvs are not updated: {','.join(sample)}")
+
+
+def wait_for_greater_than_zero_metric_value(prometheus, metric_name):
+    samples = TimeoutSampler(
+        wait_timeout=TIMEOUT_5MIN,
+        sleep=TIMEOUT_30SEC,
+        func=get_metrics_value,
+        prometheus=prometheus,
+        metrics_name=metric_name,
+    )
+    sample = None
+    try:
+        for sample in samples:
+            if sample and int(sample) > 0:
+                return
+    except TimeoutExpiredError:
+        LOGGER.info(f"Metric value of: {metric_name} is: {sample}, expected value: non zero")
+        raise
